Organize agent engine and ports
