+++ packages/agent/src/index.ts
@@ -1,15 +1,15 @@
-export { buildContext } from "./build-context";
-export { inference } from "./inference";
-export type { InferenceDeps, InferenceRequest } from "./inference";
-export { executeTools } from "./execute-tools";
+export { buildContext } from "./engine/build-context";
+export { inference } from "./engine/inference";
+export type { InferenceDeps, InferenceRequest } from "./engine/inference";
+export { executeTools } from "./engine/execute-tools";
 export type {
   ExecuteToolsDeps,
   ExecuteToolsRequest,
   ToolExecutionMode,
   ToolUpdate,
-} from "./execute-tools";
-export { nextAction } from "./next-action";
-export type { Action, RunEndStatus } from "./next-action";
-export type { Tool, ToolContext, ToolOutcome } from "./tool";
-export type { InferenceProvider, StreamRequest } from "./inference-provider";
-export type { CommitStepRequest, Store } from "./store";
+} from "./engine/execute-tools";
+export { nextAction } from "./engine/next-action";
+export type { Action, RunEndStatus } from "./engine/next-action";
+export type { Tool, ToolContext, ToolOutcome } from "./engine/tool";
+export type { InferenceProvider, StreamRequest } from "./ports/inference-provider";
+export type { CommitStepRequest, Store } from "./ports/store";
