+++ packages/agent/src/ports/fake-inference-provider.ts
